requirements.txt
pillow
flask
numpy
cv2
opencv-python